Auto boost detection in cmake
