+++ src/CMakeLists.txt
@@ -34,39 +34,62 @@ INTERFACE
 )
 add_library(WCCommon::WCCommon ALIAS WCCommon)
 
-add_library(LogConfig INTERFACE)
-target_include_directories(
-    LogConfig
-INTERFACE
-    $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/include>
-    $<INSTALL_INTERFACE:include>
-)
-target_compile_definitions(
-    LogConfig
-INTERFACE
-    SPDLOG_FMT_EXTERNAL_HO
-)
-target_link_libraries(
-    LogConfig
-INTERFACE
-    spdlog::spdlog_header_only
-)
-add_library(LogConfig::LogConfig ALIAS LogConfig)
+if(Boost_FOUND)
+    message(STATUS "Boost is found, LogConfig will be configured")
+    add_library(LogConfig INTERFACE)
+    target_include_directories(
+        LogConfig
+    INTERFACE
+        $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/include>
+        $<INSTALL_INTERFACE:include>
+    )
+    target_compile_definitions(
+        LogConfig
+    INTERFACE
+        SPDLOG_FMT_EXTERNAL_HO
+    )
+    target_link_libraries(
+        LogConfig
+    INTERFACE
+        spdlog::spdlog_header_only
+        Boost::headers
+    )
+    add_library(LogConfig::LogConfig ALIAS LogConfig)
+else()
+    message(WARNING "Boost is missing, LogConfig will not be ignored")
+endif()
 
 
 #================================================
 # Install
 #================================================
-install(TARGETS WCCommon LogConfig
-    EXPORT WCCommonTargets
-    ARCHIVE DESTINATION lib
-    INCLUDES DESTINATION include
-)
-install(DIRECTORY
-    ${CMAKE_SOURCE_DIR}/include/
-DESTINATION
-    include
-)
+if(Boost_FOUND)
+    message(STATUS "Boost is found, LogConfig will be installed")
+    install(TARGETS WCCommon LogConfig
+        EXPORT WCCommonTargets
+        ARCHIVE DESTINATION lib
+        INCLUDES DESTINATION include
+    )
+    install(DIRECTORY
+        ${CMAKE_SOURCE_DIR}/include/
+    DESTINATION
+        include
+    )
+else()
+    message(WARNING "Boost is missing, LogConfig will not be installed")
+    install(TARGETS WCCommon
+        EXPORT WCCommonTargets
+        ARCHIVE DESTINATION lib
+        INCLUDES DESTINATION include
+    )
+    install(DIRECTORY
+        ${CMAKE_SOURCE_DIR}/include/
+    DESTINATION
+        include
+    PATTERN "Def*.h" EXCLUDE
+    PATTERN "*Log*.h" EXCLUDE
+    )
+endif()
 
 set(ConfigPackageLocation lib/cmake/WCCommon)
 install(EXPORT WCCommonTargets
