+++ tests/CMakeLists.txt
@@ -17,6 +17,7 @@ find_package(Catch2 3 REQUIRED)
 # list and compile executable
 list(APPEND target_tests "NumericTimeTest")
 list(APPEND target_tests "LogConfigTest")
+list(APPEND target_tests "HJLogFormatTest")
 list(APPEND target_tests "ProgressBarTest")
 list(APPEND target_tests "CsvIOTest")
 list(APPEND target_tests "H5IOTest")
@@ -56,19 +57,19 @@ if ("NumericTimeTest" IN_LIST target_tests)
     add_test("${test_name}" ${test_name})
 endif()
 
-if ("DefEnumTest" IN_LIST target_tests)
+if (Boost_FOUND AND "DefEnumTest" IN_LIST target_tests)
     set(test_name "DefEnumTest")
     add_executable(${test_name})
     target_sources(${test_name} PUBLIC ${test_name}.cpp)
-    target_link_libraries(${test_name} PUBLIC Catch2::Catch2WithMain WCCommon::WCCommon)
+    target_link_libraries(${test_name} PUBLIC Catch2::Catch2WithMain WCCommon::WCCommon Boost::headers)
     add_test("${test_name}" ${test_name})
 endif()
 
-if ("DefTupleTest" IN_LIST target_tests)
+if (Boost_FOUND AND "DefTupleTest" IN_LIST target_tests)
     set(test_name "DefTupleTest")
     add_executable(${test_name})
     target_sources(${test_name} PUBLIC ${test_name}.cpp)
-    target_link_libraries(${test_name} PUBLIC Catch2::Catch2WithMain WCCommon::WCCommon)
+    target_link_libraries(${test_name} PUBLIC Catch2::Catch2WithMain WCCommon::WCCommon Boost::headers)
     add_test("${test_name}" ${test_name})
 endif()
 
@@ -80,8 +81,16 @@ if ("FifoFileTest" IN_LIST target_tests)
     add_test("${test_name}" ${test_name})
 endif()
 
-if ("LogConfigTest" IN_LIST target_tests)
+if (Boost_FOUND AND "LogConfigTest" IN_LIST target_tests)
     set(test_name "LogConfigTest")
+    add_executable(${test_name})
+    target_sources(${test_name} PUBLIC ${test_name}.cpp)
+    target_link_libraries(${test_name} PUBLIC Catch2::Catch2WithMain LogConfig::LogConfig WCCommon::WCCommon)
+    add_test("${test_name}" ${test_name})
+endif()
+
+if (Boost_FOUND AND "HJLogFormatTest" IN_LIST target_tests)
+    set(test_name "HJLogFormatTest")
     add_executable(${test_name})
     target_sources(${test_name} PUBLIC ${test_name}.cpp)
     target_link_libraries(${test_name} PUBLIC Catch2::Catch2WithMain LogConfig::LogConfig WCCommon::WCCommon)
@@ -112,3 +121,4 @@ if ("YAMLGetFieldTest" IN_LIST target_tests)
     add_test("${test_name}" ${test_name})
     file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/YAMLGetFieldTest.yaml DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
 endif()
+
